rename distancemaker file to entitysequenceview
