+++ vagabond/gui/AddEntity.cpp
@@ -28,6 +28,7 @@
 #include <math.h>
 #include "AddEntity.h"
 #include "ChooseEntity.h"
+#include "EntitySequenceView.h"
 #include "CalculateMetadata.h"
 
 #include <vagabond/core/Environment.h>
@@ -36,8 +37,6 @@
 #include <vagabond/core/Chain.h>
 #include <vagabond/core/Entity.h>
 
-
-#include <vagabond/gui/DistanceMaker.h>
 #include <vagabond/gui/DisplayOptions.h>
 #include <vagabond/gui/FixIssuesView.h>
 #include <vagabond/gui/ConfSpaceView.h>
