+++ vagabond/gui/EntitySequenceView.h
