+++ vagabond/gui/meson.build
@@ -44,9 +44,10 @@ guifiles = [
 'DisplayUnit.cpp',
 'DisplayUnit.h',
 'DisplayOptions.cpp',
-'DistanceMaker.cpp',
 'EntityMenu.cpp',
 'EntityFromSequence.cpp',
+'EntitySequenceView.h',
+'EntitySequenceView.cpp',
 'EvolveView.cpp',
 'EvolveView.h',
 'ExportsCSV.cpp',
@@ -157,7 +158,6 @@ guifiles = [
 'DatasetMenu.h',
 'Display.h',
 'DisplayOptions.h',
-'DistanceMaker.h',
 'EntityMenu.h',
 'EntityFromSequence.h',
 'FileLine.h',
